# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaurEye API",
    "description": "**TaurEye API** serves the Indian-equity screener, charts, and reference data\nthat power the TaurEye web and mobile apps.\n\nAll data is **factual market information only** \u2014 prices, technical indicators,\nand market caps computed from end-of-day (EOD) data. Nothing here is investment\nadvice, and TaurEye is not a SEBI-registered Investment Adviser or Research\nAnalyst.\n\n### Data freshness\nEndpoints back onto an EOD price spine. The most recent trading date is reported\nby `GET /api/indices` (`data_date`) and the universe size by `GET /api/health`.\n\n### Trying it out\nUse the **Try it out** button on any endpoint below. Request bodies (e.g.\n`POST /api/screen`) come pre-filled with a working example you can edit and run.\n\n### Authentication\nAuth is **optional** and off by default for local/LAN use. If the server sets\nthe `TAUREYE_API_KEY` environment variable (e.g. on a public host), every\n`/api/*` call except `/api/health` must present the key via the `X-API-Key`\nheader or a `?key=` query parameter. See the **ApiKeyHeader** scheme below.",
    "termsOfService": "https://taureye.app/terms",
    "contact": {
      "name": "TaurEye",
      "url": "https://taureye.app/"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8010",
      "description": "Local dev (via this machine)"
    },
    {
...
  "paths": {
...
    "/api/securities": {
      "get": {
        "tags": [
          "Securities"
        ],
        "summary": "Search securities",
        "description": "Search the universe by symbol or company name. Powers the scrip search box. Case-insensitive substring match; optionally restrict by exchange.",
        "operationId": "securities_api_securities_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search text matched against symbol and name.",
              "examples": [
                "reliance"
              ],
              "title": "Q"
            },
            "description": "Search text matched against symbol and name."
          },
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to one exchange: NSE or BSE.",
              "examples": [
                "NSE"
              ],
              "title": "Exchange"
            },
            "description": "Restrict to one exchange: NSE or BSE."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum rows to return.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum rows to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Security"
                  },
                  "title": "Response Securities Api Securities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screen": {
      "post": {
        "tags": [
          "Screener"
        ],
        "summary": "Run a screen",
        "description": "Run a filter/sort query across the entire universe and return the matching rows (capped at `limit`), the match `count`, and the `total_universe` scanned. Rows with a missing value for the sort field always sink to the bottom.",
        "operationId": "screen_api_screen_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/analysis/{symbol}": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "AI analysis (educational)",
        "description": "Educational AI summary of a scrip's factual indicators. Returns `configured=false` when no LLM key is set on the server. Always includes a disclaimer; never advice.",
        "operationId": "ai_analysis_api_ai_analysis__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Ticker symbol, e.g. RELIANCE.",
              "examples": [
                "RELIANCE"
              ],
              "title": "Symbol"
            },
            "description": "Ticker symbol, e.g. RELIANCE."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ai Analysis Api Ai Analysis  Symbol  Get"
                },
                "example": {
                  "configured": false,
                  "error": "AI analysis needs an online connection.",
                  "disclaimer": "Educational only \u2014 not investment advice. TaurEye is not a SEBI-registered Investment Adviser or Research Analyst."
                }
              }
            }
          },
          "404": {
            "description": "Unknown symbol"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/candles/{symbol}": {
      "get": {
        "tags": [
          "Securities"
        ],
        "summary": "Price history (OHLCV)",
        "description": "Daily OHLCV candles for one scrip, oldest\u2192newest, returning the last `limit` rows. Powers the chart. Prices are split/bonus-adjusted.",
        "operationId": "candles_api_candles__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Ticker symbol, e.g. RELIANCE.",
              "examples": [
                "RELIANCE"
              ],
              "title": "Symbol"
            },
            "description": "Ticker symbol, e.g. RELIANCE."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "description": "How many most-recent daily candles to return.",
              "default": 260,
              "title": "Limit"
            },
            "description": "How many most-recent daily candles to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candle"
                  },
                  "title": "Response Candles Api Candles  Symbol  Get"
                }
              }
            }
          },
          "404": {
            "description": "Unknown symbol"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screens": {
      "get": {
        "tags": [
          "Saved Screens"
        ],
        "summary": "List saved screens",
        "description": "Return all saved screen definitions on the server, newest first.",
        "operationId": "list_screens_api_screens_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedScreen"
                  },
                  "type": "array",
                  "title": "Response List Screens Api Screens Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Saved Screens"
        ],
        "summary": "Save a screen",
        "description": "Persist a named screen definition and return it with a server-assigned id and timestamp.",
        "operationId": "create_screen_api_screens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveScreenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedScreen"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/screens/{screen_id}": {
      "delete": {
        "tags": [
          "Saved Screens"
        ],
        "summary": "Delete a saved screen",
        "description": "Delete a saved screen by id. Returns the deleted id, or 404 if it does not exist.",
        "operationId": "delete_screen_api_screens__screen_id__delete",
        "parameters": [
          {
            "name": "screen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Id returned when the screen was saved.",
              "title": "Screen Id"
            },
            "description": "Id returned when the screen was saved."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Screen Api Screens  Screen Id  Delete"
                },
                "example": {
                  "deleted": "scr_a1b2c3"
                }
              }
            }
          },
          "404": {
            "description": "Screen not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Candle": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "date",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "Candle"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SavedScreen": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "request": {
            "$ref": "#/components/schemas/ScreenRequest"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "request",
          "created_at"
        ],
        "title": "SavedScreen"
      },
      "ScreenRequest": {
        "properties": {
          "filters": {
            "items": {
              "$ref": "#/components/schemas/Filter"
            },
            "type": "array",
            "title": "Filters",
            "description": "Numeric conditions to apply."
          },
          "logic": {
            "type": "string",
            "enum": [
              "AND",
              "OR"
            ],
            "title": "Logic",
            "description": "Combine filters with AND (all) or OR (any).",
            "default": "AND"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "NSE",
                  "BSE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange",
            "description": "Restrict to NSE or BSE."
          },
          "segments": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segments",
            "description": "Keep only these segments, e.g. ['EQ']. Null = all."
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Free-text match on symbol/name."
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "description": "Metric key to sort by.",
            "default": "market_cap_cr"
          },
          "sort_dir": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Sort Dir",
            "description": "Sort direction.",
            "default": "desc"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Max rows to return.",
            "default": 200
          }
        },
        "type": "object",
        "title": "ScreenRequest",
        "examples": [
          {
            "exchange": "NSE",
            "filters": [
              {
                "field": "rsi_14",
                "op": "lt",
                "value": 30
              },
              {
                "field": "close",
                "op": "gte",
                "value": 100
              }
            ],
            "limit": 50,
            "logic": "AND",
            "segments": [
              "EQ"
            ],
            "sort_by": "market_cap_cr",
            "sort_dir": "desc"
          }
        ]
      },
      "ScreenResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total_universe": {
            "type": "integer",
            "title": "Total Universe"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Metrics"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "count",
          "total_universe",
          "results"
        ],
        "title": "ScreenResponse"
      },
      "Security": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "exchange": {
            "type": "string",
            "enum": [
              "NSE",
              "BSE"
            ],
            "title": "Exchange"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "isin": {
            "type": "string",
            "title": "Isin"
          },
          "shares_outstanding": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares Outstanding"
          },
          "segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "exchange",
          "sector",
          "isin"
        ],
        "title": "Security"
      },
...
    },
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Required only when the server sets TAUREYE_API_KEY. Off for local/LAN use."
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "key",
        "description": "Alternative to the X-API-Key header (e.g. ?key=...)."
      }
    }
  },
...
}